rotation/scale sliders

+++ drawing.py
@@ -62,8 +62,8 @@ class Drawing:
         # scale
         ref_point = transform_path[0]
         transform_path = [ 
-            ((x - ref_point[0]) * scale + ref_point[0], 
-             (y - ref_point[1]) * scale + ref_point[1])
+            (int((x - ref_point[0]) * scale + ref_point[0]), 
+             int((y - ref_point[1]) * scale + ref_point[1]))
             for x, y in transform_path
         ]
         # print('scale')
